"ya funciona generalmente, roles - permisos para usuario y administrador" On branch prueba Your branch is up to date with 'origin/prueba'.
Changes to be committed:
	modified:   app/Http/Controllers/ClienteController.php
	modified:   app/Http/Controllers/PermisosController.php
	modified:   app/Http/Controllers/RoleController.php
	modified:   app/Providers/AuthServiceProvider.php
	modified:   config/adminlte.php

+++ app/Http/Controllers/PermisosController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Spatie\Permission\Models\Permission;
 
 class PermisosController extends Controller
@@ -13,6 +14,9 @@ class PermisosController extends Controller
     public function index()
     {
         //
+        if (! Gate::allows('Crear_Usuario')) {
+            abort(403);
+        }
         $permisos = Permission::all();
 
         return view('sistema.user.permisos', compact('permisos'));
@@ -58,6 +62,9 @@ class PermisosController extends Controller
     public function edit(string $id)
     {
         //
+        if (! Gate::allows('Editar_Usuario')) {
+            abort(403);
+        }
         $permisos = Permission::find($id);
 
         return view('sistema.editPermiso', compact('permisos'));
@@ -85,6 +92,9 @@ class PermisosController extends Controller
     public function destroy(string $id)
     {
         //
+        if (! Gate::allows('Eliminar_Usuario')) {
+            abort(403);
+        }
         $permisos = Permission::find($id);
         $permisos->delete();
         return back();
